Game has cors issue
