+++ client/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "axios": "^0.18.0",
     "focus-trap-react": "^5.0.1",
+    "phaser": "^3.16.2",
     "react": "^16.8.3",
     "react-dom": "^16.8.3",
     "react-router-dom": "^4.3.1",
