+++ package.json
@@ -39,6 +39,8 @@
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
     "react-bootstrap": "^1.0.0-beta.5",
-    "react-sticky": "^6.0.3"
+    "react-sticky": "^6.0.3",
+    "socket.io": "^2.2.0",
+    "socket.io-client": "^2.2.0"
   }
 }
