+++ routes/sockets.js
@@ -1,5 +1,6 @@
+
 var players = {};
-var game_config = require("../public/game_config.json");
+//var game_config = require("../public/game_config.json");
 var star = {
   x: Math.floor(Math.random() * 700) + 50,
   y: Math.floor(Math.random() * 500) + 50
@@ -8,13 +9,16 @@ var scores = {
   blue: 0,
   red: 0
 };
-var sockets = function(socket) {
+
+
+var sockets = function(socket, io) {
+
     console.log('a user connected');
     // create a new player and add it to our players object
     players[socket.id] = {
       rotation: 0,
-      x: game_config.width / 2,
-      y: game_config.height - 50,
+      x: 800 / 2,
+      y: 600 - 50,
       playerId: socket.id,
       team: (Math.floor(Math.random() * 2) == 0) ? 'red' : 'blue'
     };
